requirements.txt
google-cloud-logging==3.5.0
google-cloud-pubsub==2.18.4
